requirements-py313.txt
# Python 3.13 Compatible Requirements
# This file contains versions that are compatible with Python 3.13

# Web Framework & Database
Django==4.2.7
PyMySQL==1.1.0  # Pure Python MySQL client, no compilation needed
django-cors-headers==4.3.1
djangorestframework==3.14.0

# Quantum Computing
qiskit==0.45.0
qiskit-machine-learning==0.7.0
qiskit-aer==0.13.0

# Natural Language Processing - Updated versions for Python 3.13
spacy>=3.7.4  # Newer version with better Python 3.13 support
nltk==3.8.1
sentence-transformers>=2.3.0  # Updated for better compatibility

# Machine Learning & Data Science
scikit-learn>=1.3.2
numpy>=1.24.0,<2.0.0  # Flexible numpy version
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Task Processing & Caching
celery==5.3.4
redis==5.0.1

# Additional dependencies for better compatibility
setuptools>=65.0.0
wheel>=0.38.0
Cython>=0.29.0  # May help with compilation issues
